+++ HISTORY.md
@@ -1,7 +1,9 @@
 unreleased
 ==========
 
+  * Add `stacktrace` option
   * Add `text/plain` fallback response
+  * Remove `env` option; use `stacktrace` option instead
   * Send complete HTML document
 
 0.4.0 / 2015-06-14