+++ HISTORY.md
@@ -1,7 +1,7 @@
 unreleased
 ==========
 
-  * deps: escape-html@1.0.3
+  * deps: escape-html@~1.0.3
     - perf: enable strict mode
     - perf: optimize string replacement
     - perf: use faster string coercion